requirements.txt
Flask>=2.0
tensorflow>=2.9
numpy>=1.20
matplotlib>=3.6
Pillow>=9.0
opencv-python>=4.5.4